% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcPercentageGeneSets}
\alias{calcPercentageGeneSets}
\title{Calculate the percentage of all counts that belong to given sets of genes}
\usage{
calcPercentageGeneSets(obj, geneSets, assay = "RNA")
}
\description{
Calculate the percentage of all counts that belong to given sets of genes
}
