% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{excluFindVariableFeatures}
\alias{excluFindVariableFeatures}
\title{Find variable features while excluding given set of features using Seurat function}
\usage{
excluFindVariableFeatures(obj, exFeat, assay = "RNA", ...)
}
\description{
Find variable features while excluding given set of features using Seurat function
}
