% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeature.R
\name{getFeature}
\alias{getFeature}
\title{Get feature for 2D dimensional reduction plot}
\usage{
getFeature(
  obj,
  feat = "TOP2A",
  assayMtx = NULL,
  feat_type = c("gene", "meta"),
  assay = "RNA",
  slot = "data",
  reduc = "umap",
  dims = c(1L, 2L),
  label = T,
  label_meta = "orig.ident",
  label_max = 50L,
  vln = T,
  vln_meta = label_meta,
  vln_max = 50L,
  split = F,
  split_meta = "orig.ident",
  split_max = 10L,
  split_level = NULL,
  ...
)
}
\arguments{
\item{obj}{\emph{object} containing data needed for visualization. (Seurat object, etc.)}

\item{feat}{\emph{string} describing the feature to plot. Can be a gene or a metadata.}

\item{assayMtx}{\emph{matrix} of imputed expression value or other assay.}

\item{feat_type}{\emph{string} describing the type of feature. Can be one of:
\itemize{
\item "gene" : The feature is the name of a gene.
\item "meta" : The feature is the name of a numeric metadata.
}}

\item{assay}{\emph{string} describing which assay is used to get gene expression.
\itemize{
\item For \if{html}{\out{<u>}}Seurat\if{html}{\out{</u>}} object, "RNA" is used by default.
}}

\item{slot}{\emph{string} describing which slot of the assay is used to get gene expression.
\itemize{
\item For \if{html}{\out{<u>}}Seurat\if{html}{\out{</u>}} object, "data" is used by default.
}}

\item{reduc}{\emph{string} describing which dimension reduction is used to plot cells.}

\item{dims}{\emph{integer vector} describing which 2 dimensions are used to plot cells.}

\item{label}{\emph{logical} describing whether to label cell clusters.}

\item{label_meta}{\emph{string} describing which metadata is used to label cell clusters.}

\item{label_max}{\emph{integer} describing the maximum number of different labels.}

\item{vln}{\emph{logical} describing whether to add violin plot.}

\item{vln_meta}{\emph{string} describing which metadata is used for violin plot.}

\item{vln_max}{\emph{integer} describing the maximum number of categories of violin plot.}

\item{split}{\emph{logical} describing whether to split subplots.}

\item{split_meta}{\emph{string} describing which metadata is used for split subplots.}

\item{split_max}{\emph{integer} describing the maximum number of categories of split subplots.}

\item{split_level}{\emph{string} describing the order of categories of split subplots.}

\item{...}{Other parameters}

\item{split_nrow}{\emph{integer} describing the number of rows of split subplots if vln = F.}
}
\value{
\emph{data.table} of the feature and needed metadata
}
\description{
Get 2D dimensional reduction coordinate and the values of the feature
from given object (i.e. gene expression, score value, metadata, etc.)
}
