% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMatColor}
\alias{getMatColor}
\title{Get mild material color}
\usage{
getMatColor(pal = "blue", num = 3L)
}
\arguments{
\item{pal}{\emph{string} describing the palette used.}

\item{num}{\emph{integer} describing the number of colors.}
}
\value{
\emph{string vector} of generated colors.
}
\description{
Get mild material color
}
