% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hs2mmGene}
\alias{hs2mmGene}
\title{Transfer given human genes to mouse homologous genes}
\usage{
hs2mmGene(genes)
}
\description{
Mouse genes annotation from Cellranger mm10 v.2024-A.
Human genes annotation from Cellranger hg38 v.2024-A.
biomaRt v.2.60.1
}
