% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mm2hsGene}
\alias{mm2hsGene}
\title{Transfer given mouse genes to human homologous genes}
\usage{
mm2hsGene(genes)
}
\description{
Mouse genes annotation from Cellranger mm10 v.2024-A.
Human genes annotation from Cellranger hg38 v.2024-A.
biomaRt v.2.60.1
}
