% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeature.R
\name{plotFeature}
\alias{plotFeature}
\alias{plotFeature,Seurat-method}
\title{Better visualize feature on 2D dimensional reduction plot}
\usage{
plotFeature(obj, feat = "TOP2A", ...)

\S4method{plotFeature}{Seurat}(
  obj,
  feat = "TOP2A",
  feat_type = c("gene", "meta"),
  assay = "RNA",
  slot = "data",
  reduc = "umap",
  dims = c(1L, 2L),
  col = c("gray90", brewer.pal(9, "YlOrRd")),
  disp = seq(0, 1, len = length(col)),
  col_na = "gray90",
  cbar = c("normal", "min-max", "0-1", "pn"),
  order = c("value", "abs", "given", "random"),
  order_by = 1:ncol(obj),
  size = 0.4,
  alpha = 1,
  do_raster = F,
  dpi = 300L,
  x_name = paste0(toupper(reduc), "_", dims[1]),
  y_name = paste0(toupper(reduc), "_", dims[2]),
  label = T,
  label_meta = "orig.ident",
  label_max = 50L
)
}
\arguments{
\item{obj}{\emph{object} containing data needed for visualization. (Seurat object, etc.)}

\item{feat}{\emph{string} describing the feature to plot. Can be a gene or a numeric metadata.}

\item{...}{Other parameters}

\item{feat_type}{\emph{string} describing the type of feature. Can be one of:
\itemize{
\item "gene" : The feature is the name of a gene.
\item "meta" : The feature is the name of a numeric metadata.
}}

\item{assay}{\emph{string} describing which assay is used to get gene expression.
\itemize{
\item For \if{html}{\out{<u>}}Seurat\if{html}{\out{</u>}} object, "RNA" is used by default.
}}

\item{slot}{\emph{string} describing which slot of the assay is used to get gene expression.
\itemize{
\item For \if{html}{\out{<u>}}Seurat\if{html}{\out{</u>}} object, "data" is used by default.
}}

\item{reduc}{\emph{string} describing which dimension reduction is used to plot cells.}

\item{dims}{\emph{integer vector} describing which 2 dimensions are used to plot cells.}

\item{col}{\emph{string vector} describing the colors used to setup the continuous color bar.}

\item{disp}{\emph{numeric vector} describing the distribution pattern of the colors,
shoud have the same length as the 'col' parameter.}

\item{col_na}{\emph{string} describing the color used for NA values.}

\item{cbar}{\emph{string} describing the type of color bar. Can be one of:
\itemize{
\item "normal" : Normal color bar showing quartitles of the expression range.
\item "min-max" : Simplified color bar with only minimum and maximum labels.
\item "0-1" : Color bar forced to show values ranged from 0 to 1.
\item "pn" : Color bar showing values with both positive and negative values using
different colors.
}}

\item{order}{\emph{string} describing the way to order cells. Can be one of:
\itemize{
\item "value" : Cells will be ordered based on values of the feature.
\item "abs" : Cells will be ordered based on absolute values of the feature.
\item "given" : Cells will be ordered based on a given order.
\item "random" : Cells will be ordered randomly.
}}

\item{order_by}{\emph{integer vector} describing the order of cells when the parameter
'order' is set to "given".}

\item{size}{\emph{numeric} describing the size of cell points.}

\item{alpha}{\emph{numeric} describing the transparency of cell points.}

\item{do_raster}{\emph{logical} describing whether to perform rasterization to plots.}

\item{dpi}{\emph{integer} describing the quality of rasterization (dots per inch) when
the parameter 'do_raster' is set to TRUE.}

\item{x_name}{\emph{string} describing the name of x axis.}

\item{y_name}{\emph{string} describing the name of y axis.}

\item{label}{\emph{logical} describing whether to label cell clusters.}

\item{label_meta}{\emph{string} describing which metadata is used to label cell clusters.}

\item{label_max}{\emph{integer} describing the maximum number of different labels.}
}
\value{
\emph{ggplot object} of the plot
}
\description{
Place cells based on given 2D dimensional reduction coordinate in a specific order, and color
them using the values of a given feature (i.e. gene expression, some scores, etc.)
}
