% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMeta.R
\name{plotMeta}
\alias{plotMeta}
\alias{plotMeta,Seurat-method}
\alias{plotMeta,ArchRProject-method}
\title{Better visualize categorized metadata on 2D dimensional reduction plot}
\usage{
plotMeta(
  obj,
  meta = "orig.ident",
  meta_max = 100L,
  label = T,
  label_meta = meta,
  label_max = 50L,
  split = F,
  split_meta = meta,
  split_max = 10L,
  reduc = "umap",
  dims = c(1L, 2L),
  col = (ggsci::pal_d3("category20"))(20),
  legend_ncol = 1L,
  order = c("random", "given"),
  order_by = 1:ncol(obj),
  size = if (ncol(obj) < 5000) {
     1
 } else {
     0.4
 },
  alpha = 1,
  do_raster = F,
  dpi = 300L,
  x_name = paste0(toupper(reduc), "_", dims[1]),
  y_name = paste0(toupper(reduc), "_", dims[2]),
  ratio = 1,
  axis_tick = F,
  ...
)

\S4method{plotMeta}{Seurat}(
  obj,
  meta = "orig.ident",
  meta_max = 100L,
  label = T,
  label_meta = meta,
  label_max = 50L,
  split = F,
  split_meta = meta,
  split_max = 10L,
  reduc = "umap",
  dims = c(1L, 2L),
  col = pal_d3("category20")(20),
  legend_ncol = 1L,
  order = c("random", "given"),
  order_by = 1:ncol(obj),
  size = if (ncol(obj) < 5000) {
     1
 } else {
     0.4
 },
  alpha = 1,
  do_raster = F,
  dpi = 300L,
  x_name = paste0(toupper(reduc), "_", dims[1]),
  y_name = paste0(toupper(reduc), "_", dims[2]),
  ratio = 1,
  axis_tick = F
)

\S4method{plotMeta}{ArchRProject}(
  obj,
  meta = "Sample",
  meta_max = 100L,
  label = T,
  label_meta = meta,
  label_max = 50L,
  split = F,
  split_meta = meta,
  split_max = 10L,
  reduc = "UMAP",
  dims = c(1L, 2L),
  col = pal_d3("category20")(20),
  legend_ncol = 1L,
  order = c("random", "given"),
  order_by = 1:ncol(obj),
  size = if (length(obj$cellNames) < 5000) {
     1
 } else {
     0.4
 },
  alpha = 1,
  do_raster = F,
  dpi = 300L,
  x_name = paste0(toupper(reduc), "_", dims[1]),
  y_name = paste0(toupper(reduc), "_", dims[2]),
  ratio = 1,
  axis_tick = F
)
}
\arguments{
\item{obj}{\emph{object} containing data needed for visualization. (Seurat object, etc.)}

\item{meta}{\emph{string} describing the metadata to plot. Can be a categorized metadata.}

\item{meta_max}{\emph{integer} describing the maximum number of different categories.}

\item{label}{\emph{logical} describing whether to label cell clusters.}

\item{label_meta}{\emph{string} describing which metadata is used to label cell clusters.}

\item{label_max}{\emph{integer} describing the maximum number of different labels.}

\item{split}{\emph{logical} describing whether to split subplots.}

\item{split_meta}{\emph{string} describing which metadata is used for split subplots.}

\item{split_max}{\emph{integer} describing the maximum number of categories of split subplots.}

\item{reduc}{\emph{string} describing which dimension reduction is used to plot cells.}

\item{dims}{\emph{integer vector} describing which 2 dimensions are used to plot cells.}

\item{col}{\emph{string vector} describing the colors used to distinguish categories.}

\item{legend_ncol}{\emph{integer} describing the column number of legends.}

\item{order}{\emph{string} describing the way to order cells. Can be one of:
\itemize{
\item "random" : Cells will be ordered randomly.
\item "given" : Cells will be ordered based on a given order.
}}

\item{order_by}{\emph{integer vector} describing the order of cells when the parameter
'order' is set to "given".}

\item{size}{\emph{numeric} describing the size of cell points.}

\item{alpha}{\emph{numeric} describing the transparency of cell points.}

\item{do_raster}{\emph{logical} describing whether to perform rasterization to plot.}

\item{dpi}{\emph{integer} describing the quality of rasterization (dots per inch) when
the parameter 'do_raster' is set to TRUE.}

\item{x_name}{\emph{string} describing the name of x axis.}

\item{y_name}{\emph{string} describing the name of y axis.}

\item{ratio}{\emph{numeric} describing the axis ratio of plot.}

\item{axis_tick}{\emph{logical} describing whether to show axis ticks of plot.}

\item{...}{Other parameters}
}
\value{
\emph{ggplot object} of the plot
}
\description{
Place cells based on given 2D dimensional reduction coordinate in a specific order, and color
them using the values of a given categorized metadata (i.e. sample, celltype, etc.)
}
