% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renameAddModuleScore}
\alias{renameAddModuleScore}
\title{Calculate module scores for feature expression programs in single cells using Seurat function}
\usage{
renameAddModuleScore(obj, geneList, ...)
}
\description{
Calculate module scores for feature expression programs in single cells using Seurat function
}
