% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommend_k.R
\name{SCassist_recommend_k}
\alias{SCassist_recommend_k}
\title{Recommend `k.param` Value for Seurat's `FindNeighbors` Function}
\usage{
SCassist_recommend_k(llm_server="google", 
                 seurat_object_name, num_pcs, 
                 experimental_design = "",
                 temperature = 0,
                 max_output_tokens = 10048,
                 model_G = "gemini-1.5-flash-latest",
                 model_O = "llama3",
                 api_key_file = "api_keys.txt",
                 model_params = list(seed = 42, temperature = 0, num_gpu = 0))
}
\arguments{
\item{llm_server}{The LLM server to use. Options are "google" or "ollama". Default is "google".}

\item{seurat_object_name}{The name of the Seurat object containing the
single-cell RNA-seq data. The object should be accessible in the current
environment.}

\item{num_pcs}{The number of principal components (PCs) used for
dimensionality reduction.}

\item{experimental_design}{(Optional) A character string describing the
experimental design. This information helps the LLM contextualize the
analysis. It should include information about the sample source,
experimental conditions, research question, and clustering goals. If not
provided, the LLM will analyze the data without specific experimental
context.}

\item{temperature}{Numeric value between 0 and 1, controlling the
"creativity" of the Gemini model's response. Higher values lead to more
diverse and potentially unexpected outputs. Default is 0.}

\item{max_output_tokens}{Integer specifying the maximum number of tokens
the Gemini model can generate in its response. Default is 10048.}

\item{model_G}{Character string specifying the Gemini model to use for
analysis. Default is "gemini-1.5-flash-latest".}

\item{model_O}{Character string specifying the Ollama model to use for
analysis. Default is "llama3".}

\item{api_key_file}{Character string specifying the path to a file containing
the API key for accessing the Gemini model.}

\item{model_params}{A list of parameters to be passed to the `ollama::query` function.
This allows customization of the Llama model's behavior. Default is `list(seed = 42, temperature = 0, num_gpu = 0)`.}
}
\value{
A character string containing the LLM's recommendation for a range
 of `k.param` values, along with a justification for the choice. The
 output will be structured as:
 - "Recommended K:" followed by a range of whole number values.
 - "Reasoning:" followed by a paragraph explaining the rationale behind
 the recommendation.
}
\description{
This function analyzes key characteristics of a single-cell RNA-seq dataset
and the desired clustering goals using a large language model (LLM) to
recommend a range of potential `k.param` values for Seurat's `FindNeighbors`
function. The LLM considers factors like the number of cells, the number of
principal components (PCs) used for dimensionality reduction, and the
experimental design and clustering goals to provide a reasoned suggestion.
}
\details{
This function was written with assistance from Google's Gemini and Meta's Llama3.
}
\examples{
\dontrun{
# Assuming you have a Seurat object named 'seurat_obj'
SCassist_recommend_k(seurat_object_name = "seurat_obj",
                     num_pcs = 20,
                     experimental_design = "Single-cell RNA-seq data from a mouse brain",
                     api_key_file = "my_api_key.txt")
}
}
\author{
Vijay Nagarajan, PhD, NEI/NIH
}
